Created separate wall follower class
